+++ sdk/core/core-rest-pipeline/CHANGELOG.md
@@ -4,7 +4,7 @@
 
 ### Features Added
 
-- The `bearerTokenAuthenticationPolicy` now allows `TokenCredential` instances to configure when to begin refreshing the token by returning a `refreshOn` property in the `AccessToken` that results from the credential's `getToken` method.
+- The `bearerTokenAuthenticationPolicy` now allows `TokenCredential` instances to configure when to begin refreshing the token by returning a `refreshOnTimestamp` property in the `AccessToken` that results from the credential's `getToken` method.
 
 ### Breaking Changes